auth_service/requirements.txt
PyJWT
bcrypt
fastapi
SQLAlchemy
psycopg2-binary
pydantic[email]
uvicorn[standard]